requirement.txt
boto3
fastapi>=0.136.1
uvicorn[standard]
swagger-ui-bundle
pydantic>=2.7.0
requests>=2.32.5
langgraph
langchain-core
langchain-groq
langchain-experimental
langchain-aws
qdrant-client
pdfplumber
pytesseract
Pillow
python-dotenv
langdetect
PyYAML
